--- components/Admin/Dashboard/scss/Dashboard.module.scss
@@ -1,9 +1,0 @@
-.ant-breadcrumb {
-    background: white;
-    padding: 0px 20px;
-}
-
-.site-page-header {
-    padding: 20px 20px;
-    background: white;
-}
