+++ components/Layout/Header/Dashboard/scss/Dashboard.module.scss
@@ -1,0 +1,10 @@
+.ant-breadcrumb {
+    background: white;
+    padding-left: 20px;
+    padding-right: 20px;
+}
+
+.ant-page-layout {
+    padding: 16px 24px;
+    background-color: #fff;
+}
