+++ components/Layout/Header/Sider/data/data.js
@@ -9,7 +9,7 @@ import {
     TaiXe,
     DonHang,
     NhanVien
-} from "../../../../asstes/icons";
+} from "../../../../../asstes/icons";
 
 
 export const dataSider = [
