--- components/Layout/Header/data/dataDropdown.js
@@ -1,4320 +1,0 @@
-export const dataDropdown = {
-    "search": {
-        "totalCount": 7,
-        "edges": [
-            {
-                "node": {
-                    "id": "Q2F0ZWdvcnk6Mjg2",
-                    "name": "Dược phẩm",
-                    "slug": "dược-phẩm",
-                    "level": 0,
-                    "icon": {
-                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/01_D%C6%B0%E1%BB%A3c_ph%E1%BA%A9m_01-Medicine__Duoc_Pham.png",
-                        "alt": ""
-                    },
-                    "backgroundImage": null,
-                    "translation": null,
-                    "metadata": [
-                        {
-                            "key": "color",
-                            "value": "#5DAC45"
-                        }
-                    ],
-                    "parent": null,
-                    "children": {
-                        "edges": [
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6Mjg3",
-                                    "name": "Thuốc không kê đơn",
-                                    "slug": "thuốc-không-kê-đơn",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/Thu%E1%BB%91c_ko_k%C3%AA_%C4%91%C6%A1n.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [
-                                        {
-                                            "key": "requirePharmacist",
-                                            "value": "1"
-                                        }
-                                    ],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjg4",
-                                                    "name": "Thuốc kháng dị ứng",
-                                                    "slug": "thuốc-kháng-dị-ứng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.01_OTC_Thu%E1%BB%91c_kh%C3%A1ng_d%E1%BB%8B_%E1%BB%A9ng_1.01.01-OTC_Antiallergics_OTC_Thuoc_khang_di_ung.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjg5",
-                                                    "name": "Thuốc kháng viêm",
-                                                    "slug": "thuốc-kháng-viêm",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.02_OTC_Thu%E1%BB%91c_kh%C3%A1ng_vi%C3%AAm_1.01.02-OTC_Anti-flammatory_OTC_Thuoc_khang_viem_2.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjkw",
-                                                    "name": "Thuốc ngừa thai",
-                                                    "slug": "thuốc-ngừa-thai",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.03_OTC_Thu%E1%BB%91c_ng%E1%BB%ABa_thai_1.01.03-OTC_Contraceptives_OTC_Thuoc_ngua_thai.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjkx",
-                                                    "name": "Thuốc cảm lạnh",
-                                                    "slug": "thuốc-cảm-lạnh",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.04_OTC_Thu%E1%BB%91c_c%E1%BA%A3m_l%E1%BA%A1nh_ho_1.01.04-OTC_Cough__Cold_OTC_Thuoc_cam_lanh_ho.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjky",
-                                                    "name": "Thuốc da liễu",
-                                                    "slug": "thuốc-da-liễu",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.05_OTC_Thu%E1%BB%91c_da_li%E1%BB%85u_1.01.05-OTC_Dermatological_OTC_Thuoc_da_lieu.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjkz",
-                                                    "name": "Thuốc giảm cân",
-                                                    "slug": "thuốc-giảm-cân",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.06_OTC_Thu%E1%BB%91c_gi%E1%BA%A3m_c%C3%A2n_1.01.06-OTC_Dietery_OTC_Thuoc_giam_can.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjk0",
-                                                    "name": "Thuốc Mắt/Tai/Mũi",
-                                                    "slug": "thuốc-mắttaimũi",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.07_OTC_Thu%E1%BB%91c_M%E1%BA%AFtTaiM%C5%A9i_1.01.07-OTC_EyeEarNose_OTC_Thuoc_MatTaiMui_2.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjk1",
-                                                    "name": "Thuốc tiêu hóa",
-                                                    "slug": "thuốc-tiêu-hóa",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.08_OTC_Thu%E1%BB%91c_ti%C3%AAu_h%C3%B3a_1.01.08-OTC_Gastrointestinal_OTC_Thuoc_tieu_hoa_2.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjk2",
-                                                    "name": "Thuốc dành cho nam",
-                                                    "slug": "thuốc-dành-cho-nam",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.09_OTC_Thu%E1%BB%91c_d%C3%A0nh_cho_nam_1.01.09-OTC_Mens_Health_OTC_Thuoc_danh_cho_nam1.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjk3",
-                                                    "name": "Giảm đau, hạ sốt",
-                                                    "slug": "giảm-đau-hạ-sốt",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.10_OTC_Gi%E1%BA%A3m_%C4%91au_h%E1%BA%A1_s%E1%BB%91t_1.01.10-OTC_Pain__Fever_Relief_OTC_Giam_dau_ha_sot_co.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjk4",
-                                                    "name": "Thuốc dành cho phụ nữ",
-                                                    "slug": "thuốc-dành-cho-phụ-nữ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.11_OTC_Thu%E1%BB%91c_d%C3%A0nh_cho_ph%E1%BB%A5_n%E1%BB%AF_1.01.11-OTC_Womens_Health_OTC_Thuoc_danh_cho_ph.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mjk5",
-                                                    "name": "Thuốc thần kinh",
-                                                    "slug": "thuốc-thần-kinh",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.12_OTC_Thu%E1%BB%91c_th%E1%BA%A7n_kinh_1.01.12-OTC_Neurotrophics_OTC_Thuoc_than_kinh.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzAw",
-                                                    "name": "Thuốc cơ xương khớp",
-                                                    "slug": "thuốc-cơ-xương-khớp",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.13_OTC_Thu%E1%BB%91c_c%C6%A1_x%C6%B0%C6%A1ng_kh%E1%BB%9Bp_1.01.13-OTC_Musculo-Skeletal_System_OTC_Thuoc_co_.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzAx",
-                                                    "name": "Vitamin & Khoáng chất",
-                                                    "slug": "vitamin-khoáng-chất",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.14_OTC_Vitamin__Kho%C3%A1ng_ch%E1%BA%A5t_1.01.14-OTC_Vitamins__Minerals_OTC_Vitamin__Khoa.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzAy",
-                                                    "name": "Dầu gió, dầu cù là ...",
-                                                    "slug": "dầu-gió-dầu-cù-là",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.15_OTC_D%E1%BA%A7u_gi%C3%B3_d%E1%BA%A7u_c%C3%B9_l%C3%A0_..._1.01.15-OTC_Medicated_Oil__Ointment_Dau_gio_dau.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzAz",
-                                                    "name": "Thuốc khác",
-                                                    "slug": "thuốc-khác",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.01.16_OTC_Thu%E1%BB%91c_kh%C3%A1c_1.01.16-OTC_Others_OTC_Thuoc_khac1.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTU0",
-                                    "name": "Thuốc kê đơn",
-                                    "slug": "thuốc-kê-đơn",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/1.02_Thu%E1%BB%91c_k%C3%AA_%C4%91%C6%A1n_1.02-Rx_Thuoc_ke_don.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [
-                                        {
-                                            "key": "recipe",
-                                            "value": "1"
-                                        }
-                                    ],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTc1",
-                                                    "name": "RX Sản phẩm khác",
-                                                    "slug": "rx-sản-phẩm-khác",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTc2",
-                                                    "name": "RX Kháng dị ứng",
-                                                    "slug": "rx-kháng-dị-ứng",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTc3",
-                                                    "name": "Rx Thuốc tiểu đường",
-                                                    "slug": "rx-thuốc-tiểu-đường",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTc4",
-                                                    "name": "Rx Thuốc kháng viêm",
-                                                    "slug": "rx-thuốc-kháng-viêm",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTc5",
-                                                    "name": "Rx Thuốc kháng sinh",
-                                                    "slug": "rx-thuốc-kháng-sinh",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTgw",
-                                                    "name": "Rx Thuốc tim mạch, huyết áp",
-                                                    "slug": "rx-thuốc-tim-mạch-huyết-áp",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTgx",
-                                                    "name": "Rx Thuốc ngừa thai",
-                                                    "slug": "rx-thuốc-ngừa-thai",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTgy",
-                                                    "name": "Rx Thuốc cảm lạnh, ho",
-                                                    "slug": "rx-thuốc-cảm-lạnh-ho",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTgz",
-                                                    "name": "Rx Thuốc da liễu",
-                                                    "slug": "rx-thuốc-da-liễu",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTg0",
-                                                    "name": "Rx Thuốc Mắt/Tai/Mũi",
-                                                    "slug": "rx-thuốc-mắttaimũi",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTg1",
-                                                    "name": "Rx Thuốc tiêu hóa",
-                                                    "slug": "rx-thuốc-tiêu-hóa",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTg2",
-                                                    "name": "Rx Thuốc cho nam giới",
-                                                    "slug": "rx-thuốc-cho-nam-giới",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTg3",
-                                                    "name": "Rx Thuốc giảm đau, hạ sốt",
-                                                    "slug": "rx-thuốc-giảm-đau-hạ-sốt",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTg4",
-                                                    "name": "Rx Thuốc dành cho nữ",
-                                                    "slug": "rx-thuốc-dành-cho-nữ",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTg5",
-                                                    "name": "Rx Thuốc thần kinh",
-                                                    "slug": "rx-thuốc-thần-kinh",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTkw",
-                                                    "name": "Rx Thuốc cơ xương khớp",
-                                                    "slug": "rx-thuốc-cơ-xương-khớp",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTkx",
-                                                    "name": "Rx Vitamin & Khoáng chất",
-                                                    "slug": "rx-vitamin-khoáng-chất",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTky",
-                                                    "name": "Rx Hệ hô hấp",
-                                                    "slug": "rx-hệ-hô-hấp",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTkz",
-                                                    "name": "Rx Thuốc khác",
-                                                    "slug": "rx-thuốc-khác",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            },
-            {
-                "node": {
-                    "id": "Q2F0ZWdvcnk6MzI1",
-                    "name": "Chăm sóc sức khỏe",
-                    "slug": "chăm-sóc-sức-khỏe",
-                    "level": 0,
-                    "icon": {
-                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/02_Ch%C4%83m_s%C3%B3c_s%E1%BB%A9c_kh%E1%BB%8Fe_02-Healthcare_Cham_soc_suc_khoe.png",
-                        "alt": ""
-                    },
-                    "backgroundImage": null,
-                    "translation": null,
-                    "metadata": [
-                        {
-                            "key": "color",
-                            "value": "#834198"
-                        }
-                    ],
-                    "parent": null,
-                    "children": {
-                        "edges": [
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6MzI2",
-                                    "name": "Dược mỹ phẩm",
-                                    "slug": "dược-mỹ-phẩm",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.01_D%C6%B0%E1%BB%A3c_m%E1%BB%B9_ph%E1%BA%A9m_2.01-Dermo_Skincare_Duoc_my_pham1.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzI3",
-                                                    "name": "Tinh chất & Serum",
-                                                    "slug": "tinh-chất-serum",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.01.01_Tinh_ch%E1%BA%A5t__Serum_2.01.01-Dermo_Essence__Serum_Tinh_chat__Serum.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzI4",
-                                                    "name": "Kem dưỡng ẩm",
-                                                    "slug": "kem-dưỡng-ẩm",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.01.02_Kem_d%C6%B0%E1%BB%A1ng_%E1%BA%A9m_2.01.02-Dermo_Moisturizers__Cream_Kem_duong_am.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzI5",
-                                                    "name": "Nước cân bằng & Xịt khoáng",
-                                                    "slug": "nước-cân-bằng-xịt-khoáng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.01.03_N%C6%B0%E1%BB%9Bc_c%C3%A2n_b%E1%BA%B1ng__X%E1%BB%8Bt_kho%C3%A1ng_2.01.03-Dermo_Toner__Mist_Nuoc_can_bang__xit_kh.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzMw",
-                                                    "name": "Sản phẩm trị thẹo",
-                                                    "slug": "sản-phẩm-trị-thẹo",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.01.07_S%E1%BA%A3n_ph%E1%BA%A9m_tr%E1%BB%8B_th%E1%BA%B9o_2.01.07-Scar_treatment_San_pham_tri_theo_.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzMx",
-                                                    "name": "Sản phẩm chống lão hóa",
-                                                    "slug": "sản-phẩm-chống-lão-hóa",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.01.14_S%E1%BA%A3n_ph%E1%BA%A9m_ch%E1%BB%91ng_l%C3%A3o_h%C3%B3a_2.01.14-Anti-aging_San_pham_chong_lao_hoa.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzMy",
-                                                    "name": "Sản phẩm trị mụn",
-                                                    "slug": "sản-phẩm-trị-mụn",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.01.15_S%E1%BA%A3n_ph%E1%BA%A9m_tr%E1%BB%8B_m%E1%BB%A5n_2.01.15-Acnes_Treatment_San_pham_tri_mun.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6MzMz",
-                                    "name": "Thực phẩm dinh dưỡng",
-                                    "slug": "thực-phẩm-dinh-dưỡng",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.04Th%E1%BB%B1c_ph%E1%BA%A9m_dinh_d%C6%B0%E1%BB%A1ng_2.04-Nutrition_Food_Thuc_pham_dinh_duong.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzM0",
-                                                    "name": "Sữa dinh dưỡng",
-                                                    "slug": "sữa-dinh-dưỡng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.04.01_S%E1%BB%AFa_dinh_d%C6%B0%E1%BB%A1ng_2.04.01-Milk_Powder_Sua_bot.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzM1",
-                                                    "name": "Thức uống có lợi cho sức khỏe",
-                                                    "slug": "thức-uống-có-lợi-cho-sức-khỏe",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzM2",
-                                                    "name": "Thực phẩm có lợi cho sức khỏe",
-                                                    "slug": "thực-phẩm-có-lợi-cho-sức-khỏe",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.04.03_Th%E1%BB%B1c_ph%E1%BA%A9m_c%C3%B3_l%E1%BB%A3i_cho_s%E1%BB%A9c_kh%E1%BB%8Fe_2.04.03-Healthy_Food_Thuc_pham_co_loi_cho_s.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzM3",
-                                                    "name": "Dinh dưỡng thể thao-thể hình",
-                                                    "slug": "dinh-dưỡng-thể-thao-thể-hình",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.04.04_Dinh_d%C6%B0%E1%BB%A1ng_th%E1%BB%83_thao-th%E1%BB%83_h%C3%ACnh.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6MzM4",
-                                    "name": "Thiết bị y tế",
-                                    "slug": "thiết-bị-y-tế",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.05_Thi%E1%BA%BFt_b%E1%BB%8B_y_t%E1%BA%BF_2.05-Healthcare_Equipment_Thiet_bi_y_te.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzM5",
-                                                    "name": "Nhiệt kế",
-                                                    "slug": "nhiệt-kế",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.05.01_Nhi%E1%BB%87t_k%E1%BA%BF_2.05.01-Thermometers_Nhiet_ke_l.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzQw",
-                                                    "name": "Máy đo huyết áp",
-                                                    "slug": "máy-đo-huyết-áp",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.05.02_M%C3%A1y_%C4%91o_huy%E1%BA%BFt_%C3%A1p_2.05.02-Blood_Pressure_Monitors_May_do_huyet_ap.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzQx",
-                                                    "name": "Máy đo đường huyết",
-                                                    "slug": "máy-đo-đường-huyết",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.05.03_M%C3%A1y_%C4%91o_%C4%91%C6%B0%E1%BB%9Dng_huy%E1%BA%BFt_2.05.03-Blood_Glucose_Monitors_May_do_duong_huyet.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzQy",
-                                                    "name": "Máy xông khí dung",
-                                                    "slug": "máy-xông-khí-dung",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.05.04_M%C3%A1y_x%C3%B4ng_kh%C3%AD_dung_2.05.04-Nebulizer__Aspirators_May_xong_khi_dung.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzQz",
-                                                    "name": "Thiết bị y tế khác",
-                                                    "slug": "thiết-bị-y-tế-khác",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.05.05_Thi%E1%BA%BFt_b%E1%BB%8B_y_t%E1%BA%BF_kh%C3%A1c_2.05.05-Other_Medical_Devices_Thiet_bi_y_te_khac.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6MzQ0",
-                                    "name": "Dụng cụ sơ cứu",
-                                    "slug": "dụng-cụ-sơ-cứu",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.06_D%E1%BB%A5ng_c%E1%BB%A5_s%C6%A1_c%E1%BB%A9u_2.06-First_Aids_Dung_cu_so_cuu1.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzQ2",
-                                                    "name": "Băng gạc",
-                                                    "slug": "băng-gạc",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.06.01_B%C4%83ng_g%E1%BA%A1c_2.06.01-Bandages_Bang_gac.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzQ3",
-                                                    "name": "Sản phẩm chăm sóc vết thương",
-                                                    "slug": "sản-phẩm-chăm-sóc-vết-thương",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.06.02_S%E1%BA%A3n_ph%E1%BA%A9m_ch%C4%83m_s%C3%B3c_v%E1%BA%BFt_th%C6%B0%C6%A1ng_2.06.02-General_Wound_Care_Treatment_San_pha.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzQ4",
-                                                    "name": "Hỗ trợ chấn thương thể thao",
-                                                    "slug": "hỗ-trợ-chấn-thương-thể-thao",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.06.03_H%E1%BB%97_tr%E1%BB%A3_ch%E1%BA%A5n_th%C6%B0%C6%A1ng_th%E1%BB%83_thao_2.06.03-Sport_Injury_Support_Ho_tro_chan_thuo.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzQ5",
-                                                    "name": "Vớ chống giãn tĩnh mạch",
-                                                    "slug": "vớ-chống-giãn-tĩnh-mạch",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.06.04_V%E1%BB%9B_ch%E1%BB%91ng_gi%C3%A3n_t%C4%A9nh_m%E1%BA%A1ch_2.06.04-Compression_Stockings_Vo_chong_gian_tinh_.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzUw",
-                                                    "name": "Dụng cụ thể thao",
-                                                    "slug": "sport-equipment-dung-cu-the-thao",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6MzUx",
-                                    "name": "Kế hoạch gia đình",
-                                    "slug": "kế-hoạch-gia-đình",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.07_K%E1%BA%BF_ho%E1%BA%A1ch_gia_%C4%91%C3%ACnh_2.07-Family_Planning_Ke_hoach_gia_dinh_-_Copy.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzUy",
-                                                    "name": "Bao cao su",
-                                                    "slug": "bao-cao-su",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.07.01_Bao_cao_su_2.07.01-Condoms_Bao_cao_su.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzUz",
-                                                    "name": "Gel bôi trơn",
-                                                    "slug": "gel-bôi-trơn",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.07.02_Gel_b%C3%B4i_tr%C6%A1n_2.07.02-Lubricating_Gel_Gel_boi_tron.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzU0",
-                                                    "name": "Que thử thai",
-                                                    "slug": "que-thử-thai",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.07.03_Que_th%E1%BB%AD_thai_2.07.03-Pregnancy_Test_Que_thu_thai_2.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzU1",
-                                                    "name": "Que thử rụng trứng",
-                                                    "slug": "que-thử-rụng-trứng",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-backgrounds/2.07.04_Que_th%E1%BB%AD_r%E1%BB%A5ng_tr%E1%BB%A9ng_2.07.04-Ovulation_Test_Que_thu_rung_trung.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6MzU2",
-                                    "name": "Chăm sóc Mắt/Tai/Mũi",
-                                    "slug": "chăm-sóc-mắttaimũi",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08_Ch%C4%83m_s%C3%B3c_M%E1%BA%AFtTaiM%C5%A9i_2.08-For_EyeEarNose_Cham_soc_MatTaiMui.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzU3",
-                                                    "name": "Kính sát tròng",
-                                                    "slug": "kính-sát-tròng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.01_K%C3%ADnh_s%C3%A1t_tr%C3%B2ng_2.08.01-Contact_Lens_Kinh_sat_trong.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzU4",
-                                                    "name": "Nước ngâm kính sát tròng",
-                                                    "slug": "nước-ngâm-kính-sát-tròng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.02_N%C6%B0%E1%BB%9Bc_ng%C3%A2m_k%C3%ADnh_s%C3%A1t_tr%C3%B2ng_2.08.02-Contact_Lens_Solution_Nuoc_ngam_kinh_sat.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzU5",
-                                                    "name": "Thuốc nhỏ mắt",
-                                                    "slug": "thuốc-nhỏ-mắt",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.03_Thu%E1%BB%91c_nh%E1%BB%8F_m%E1%BA%AFt_2.08.03-Eyedrop_Thuoc_nho_mat.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzYw",
-                                                    "name": "Mỡ mắt và Thuốc nhỏ mắt chuyên sâu",
-                                                    "slug": "mỡ-mắt-và-thuốc-nhỏ-mắt-chuyên-sâu",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzYx",
-                                                    "name": "Thuốc nhỏ tai",
-                                                    "slug": "thuốc-nhỏ-tai",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.05_Thu%E1%BB%91c_nh%E1%BB%8F_tai_2.08.05-Eardrop_Thuoc_nho_tai.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzYy",
-                                                    "name": "Thuốc xịt mũi",
-                                                    "slug": "thuốc-xịt-mũi",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.06_Thu%E1%BB%91c_x%E1%BB%8Bt_m%C5%A9i_2.08.06-Nasal_Spray_Thuoc_xit_mui.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzYz",
-                                                    "name": "Sản phẩm khác",
-                                                    "slug": "sản-phẩm-khác-2",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.07_S%E1%BA%A3n_ph%E1%BA%A9m_kh%C3%A1c_2.08.07-Others_San_pham_khac.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzY0",
-                                                    "name": "Thuốc xịt mũi trẻ em",
-                                                    "slug": "nasal-spray-for-kids-thuoc-xit-mui-tre-em",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.08_2.08.08-Nasal_Spray_For_Kids_Thuoc_xit_mui_tre_em_2.08.08-Nasal_Spray_For.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzY1",
-                                                    "name": "Kính mắt",
-                                                    "slug": "sun-glasses-kinh-mat",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.09_2.08.09-Sun_Glasses_Kinh_Mat_2.08.09-Sun_Glasses_Kinh_Mat.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzY2",
-                                                    "name": "Kính đọc",
-                                                    "slug": "reading-glasses-kinh-doc",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.10_2.08.10-Reading_Glasses_Kinh_Doc_2.08.10-Reading_Glasses_Kinh_Doc.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzY3",
-                                                    "name": "Kính bảo hộ",
-                                                    "slug": "safety-glasses-kinh-bao-ho",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.08.11_2.08.11-Safety_Glasses.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6MzY4",
-                                                    "name": "Kính bơi",
-                                                    "slug": "swimming-glasses-kinh-boi",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6MzY5",
-                                    "name": "Chăm sóc chân",
-                                    "slug": "chăm-sóc-chân",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.09_Ch%C4%83m_s%C3%B3c_ch%C3%A2n_2.09-Foot_Care_Cham_soc_chan2.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzcw",
-                                                    "name": "Miếng đệm chân",
-                                                    "slug": "miếng-đệm-chân",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzcx",
-                                                    "name": "Sản phẩm khử mùi hôi chân",
-                                                    "slug": "sản-phẩm-khử-mùi-hôi-chân",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.09.02_S%E1%BA%A3n_ph%E1%BA%A9m_kh%E1%BB%AD_m%C3%B9i_h%C3%B4i_ch%C3%A2n_2.09.02-Foot_Deodorant_San_pham_khu_mui_hoi_cha.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzcy",
-                                                    "name": "Dụng cụ chăm sóc chân",
-                                                    "slug": "dụng-cụ-chăm-sóc-chân",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.09.03_D%E1%BB%A5ng_c%E1%BB%A5_ch%C4%83m_s%C3%B3c_ch%C3%A2n_2.09.03-Foot_Accessories_Dung_cu_cham_soc_chan.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzcz",
-                                                    "name": "Kem trị nứt gót chân",
-                                                    "slug": "kem-trị-nứt-gót-chân",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.09.04_Kem_tr%E1%BB%8B_n%E1%BB%A9t_g%C3%B3t_ch%C3%A2n_2.09.04-Heel_Cracker_Cream_Kem_tri_nut_got_chan.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzc0",
-                                                    "name": "Dưỡng da chân & tẩy tế bào chết",
-                                                    "slug": "dưỡng-da-chân-tẩy-tế-bào-chết",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzc1",
-                                                    "name": "Mặt nạ chân",
-                                                    "slug": "mặt-nạ-chân",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6Mzc2",
-                                    "name": " Khẩu trang y tế",
-                                    "slug": "khẩu-trang-y-tế",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.10_Kh%E1%BA%A9u_trang_y_t%E1%BA%BF_2.10-Medical_Face_Mask_Khau_trang_y_te.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": []
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6Mzc3",
-                                    "name": "Chống muỗi",
-                                    "slug": "chống-muỗi",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.11.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": []
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6Mzc4",
-                                    "name": "Dầu tràm, dầu xoa bóp",
-                                    "slug": "dầu-tràm-dầu-xoa-bóp",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/2.12.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": []
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            },
-            {
-                "node": {
-                    "id": "Q2F0ZWdvcnk6Mzc5",
-                    "name": "Chăm sóc cá nhân",
-                    "slug": "chăm-sóc-cá-nhân",
-                    "level": 0,
-                    "icon": {
-                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/03_Ch%C4%83m_s%C3%B3c_c%C3%A1_nh%C3%A2n_03-Personal_Care_Cham_soc_ca_nhan.png",
-                        "alt": ""
-                    },
-                    "backgroundImage": null,
-                    "translation": null,
-                    "metadata": [
-                        {
-                            "key": "color",
-                            "value": "#0D4DA1"
-                        }
-                    ],
-                    "parent": null,
-                    "children": {
-                        "edges": [
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6Mzgw",
-                                    "name": "Sản phẩm phòng tắm",
-                                    "slug": "sản-phẩm-phòng-tắm",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.02_S%E1%BA%A3n_ph%E1%BA%A9m_ph%C3%B2ng_t%E1%BA%AFm_3.02-Bath_Care_San_pham_phong_tam.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzgx",
-                                                    "name": "Sữa tắm",
-                                                    "slug": "sữa-tắm",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.02.01_S%E1%BB%AFa_t%E1%BA%AFm_3.02.01-Shower_Gel_Sua_tam.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzgy",
-                                                    "name": "Xà bông cục",
-                                                    "slug": "xà-bông-cục",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.02.02_X%C3%A0_b%C3%B4ng_c%E1%BB%A5c_3.02.02-Soap_Xa_bong_cuc.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzgz",
-                                                    "name": "Phụ kiện phòng tắm",
-                                                    "slug": "phụ-kiện-phòng-tắm",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.02.03_Ph%E1%BB%A5_ki%E1%BB%87n_ph%C3%B2ng_t%E1%BA%AFm_3.02.03-Bath_accessories_Phu_kien_phong_tam.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6Mzg0",
-                                    "name": "Chăm sóc cơ thể",
-                                    "slug": "chăm-sóc-cơ-thể",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.03_Ch%C4%83m_s%C3%B3c_c%C6%A1_th%E1%BB%83_3.03-Body_Care_Cham_soc_co_the.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzg1",
-                                                    "name": "Dưỡng thể",
-                                                    "slug": "dưỡng-thể",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.03.01_D%C6%B0%E1%BB%A1ng_th%E1%BB%83_3.03.01-Body_Lotion_Kem_duong_the.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzg2",
-                                                    "name": "Dầu dưỡng thể & massage",
-                                                    "slug": "dầu-dưỡng-thể-massage",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.03.02_D%E1%BA%A7u_d%C6%B0%E1%BB%A1ng_th%E1%BB%83__massage_3.03.02-Body__Massage_Oil_Dau_duong_the__massage.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzg3",
-                                                    "name": "Tẩy tế bào chết toàn thân",
-                                                    "slug": "tẩy-tế-bào-chết-toàn-thân",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.03.03_T%E1%BA%A9y_t%E1%BA%BF_b%C3%A0o_ch%E1%BA%BFt_to%C3%A0n_th%C3%A2n_3.03.03-Body_Scrub_Tay_te_bao_chet_toan_than.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzg4",
-                                                    "name": "Kem chống nắng toàn thân",
-                                                    "slug": "kem-chống-nắng-toàn-thân",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.03.04_Kem_ch%E1%BB%91ng_n%E1%BA%AFng_to%C3%A0n_th%C3%A2n_3.03.04-Suncare_for_body_Kem_chong_nang_toan_tha.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzg5",
-                                                    "name": "Dưỡng da tay & tẩy tế bào chết",
-                                                    "slug": "dưỡng-da-tay-tẩy-tế-bào-chết",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzkw",
-                                                    "name": "Nước rửa tay",
-                                                    "slug": "nước-rửa-tay",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.03.06_N%C6%B0%E1%BB%9Bc_r%E1%BB%ADa_tay_3.03.06-Hand_wash__sanitizers_Nuoc_rua_tay.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzkx",
-                                                    "name": "Mặt nạ tay",
-                                                    "slug": "mặt-nạ-tay",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6Mzky",
-                                    "name": "Sản phẩm khử mùi",
-                                    "slug": "sản-phẩm-khử-mùi",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.04_S%E1%BA%A3n_ph%E1%BA%A9m_kh%E1%BB%AD_m%C3%B9i_3.04-Deodorant_San_pham_khu_mui.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzkz",
-                                                    "name": "Lăn khử mùi",
-                                                    "slug": "lăn-khử-mùi",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.04.01_L%C4%83n_kh%E1%BB%AD_m%C3%B9i_3.04.01-Roll_Lan_khu_mui.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzk0",
-                                                    "name": "Xịt khử mùi",
-                                                    "slug": "xịt-khử-mùi",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.04.02_X%E1%BB%8Bt_kh%E1%BB%AD_m%C3%B9i_3.04.02-Spray_Xit_khu_mui.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6Mzk1",
-                                                    "name": "Sáp khử mùi",
-                                                    "slug": "sáp-khử-mùi",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.04.03_S%C3%A1p_kh%E1%BB%AD_m%C3%B9i_3.04.03-Stick_Sap_khu_mui.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6Mzk3",
-                                    "name": "Chăm sóc tóc",
-                                    "slug": "chăm-sóc-tóc",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.05_Ch%C4%83m_s%C3%B3c_t%C3%B3c_3.05-Hair_Care_Cham_soc_toc.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDAy",
-                                                    "name": "Dầu gội đầu",
-                                                    "slug": "dầu-gội-đầu",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.05.01_D%E1%BA%A7u_g%E1%BB%99i_%C4%91%E1%BA%A7u_3.05.01-Shampoo_Dau_goi_dau.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDAz",
-                                                    "name": "Dầu xả",
-                                                    "slug": "dầu-xả",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.05.02_D%E1%BA%A7u_x%E1%BA%A3_3.05.02-Conditioner_Dau_xa.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDA0",
-                                                    "name": "Dưỡng tóc",
-                                                    "slug": "dưỡng-tóc",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.05.03_D%C6%B0%E1%BB%A1ng_t%C3%B3c_3.05.03-Hair_Treatment_Duong_toc.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDA1",
-                                                    "name": "Tạo kiểu tóc",
-                                                    "slug": "tạo-kiểu-tóc",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.05.04_T%E1%BA%A1o_ki%E1%BB%83u_t%C3%B3c_3.05.04-Hair_Styling_Tao_kieu_toc.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDA2",
-                                                    "name": "Nhuộm tóc",
-                                                    "slug": "nhuộm-tóc",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.05.05_Nhu%E1%BB%99m_t%C3%B3c_3.05.05-Hair_Coloring_Nhuom_toc.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDA3",
-                                    "name": "Vệ sinh phụ nữ",
-                                    "slug": "vệ-sinh-phụ-nữ",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.06_V%E1%BB%87_sinh_ph%E1%BB%A5_n%E1%BB%AF_3.06-Feminine_Hygiene_Ve_sinh_phu_nu.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDA4",
-                                                    "name": "Dung dịch vệ sinh phụ nữ",
-                                                    "slug": "dung-dịch-vệ-sinh-phụ-nữ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.06.01_Dung_d%E1%BB%8Bch_v%E1%BB%87_sinh_ph%E1%BB%A5_n%E1%BB%AF_3.06.01-Feminine_Wash_Dung_dich_ve_sinh_phu_nu.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDA5",
-                                                    "name": "BVS hàng ngày",
-                                                    "slug": "bvs-hàng-ngày",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.06.02_BVS_h%C3%A0ng_ng%C3%A0y_3.06.02-Panty_liners_BVS_hang_ngay.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDEw",
-                                                    "name": "Băng vệ sinh",
-                                                    "slug": "băng-vệ-sinh",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.06.03_B%C4%83ng_v%E1%BB%87_sinh_3.06.03-Sanitary_Napkins_Bang_ve_sinh.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDEx",
-                                                    "name": "BVS dạng ống",
-                                                    "slug": "bvs-dạng-ống",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.06.04_BVS_d%E1%BA%A1ng_%E1%BB%91ng_3.06.04-Tampons_BVS_dang_ong.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDEy",
-                                                    "name": "Sản phẩm vệ sinh khác",
-                                                    "slug": "sản-phẩm-vệ-sinh-khác",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.06.05_S%E1%BA%A3n_ph%E1%BA%A9m_v%E1%BB%87_sinh_kh%C3%A1c_3.06.05-Others_San_pham_ve_sinh_khac.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDEz",
-                                    "name": "Chăm sóc nam giới",
-                                    "slug": "chăm-sóc-nam-giới",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.07_Ch%C4%83m_s%C3%B3c_nam_gi%E1%BB%9Bi_3.07-Men_Care_Cham_soc_nam_gioi.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDE0",
-                                                    "name": "Dao cạo râu & Bọt cạo râu",
-                                                    "slug": "dao-cạo-râu-bọt-cạo-râu",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.07.01_Dao_c%E1%BA%A1o_r%C3%A2u__B%E1%BB%8Dt_c%E1%BA%A1o_r%C3%A2u_3.07.01-Shaving__Grooming_Dao_cao_rau__Bot_cao_r.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDE1",
-                                                    "name": "Chăm sóc tóc cho nam",
-                                                    "slug": "chăm-sóc-tóc-cho-nam",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.07.02_Ch%C4%83m_s%C3%B3c_t%C3%B3c_cho_nam_3.07.02-Hair_Care_Cham_soc_toc_cho_nam.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDE2",
-                                                    "name": "Chăm sóc da mặt cho nam",
-                                                    "slug": "chăm-sóc-da-mặt-cho-nam",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.07.03_Ch%C4%83m_s%C3%B3c_da_m%E1%BA%B7t_cho_nam_3.07.03-Facial_Care_Cham_soc_da_mat_cho_nam.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDE3",
-                                                    "name": "Sản phẩm tắm & dưỡng thể cho nam",
-                                                    "slug": "sản-phẩm-tắm-dưỡng-thể-cho-nam",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.07.04_S%E1%BA%A3n_ph%E1%BA%A9m_t%E1%BA%AFm__d%C6%B0%E1%BB%A1ng_th%E1%BB%83_cho_nam_3.07.04-Bath__Body_Care_San_pham_tam__duo.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDE4",
-                                                    "name": "Khử mùi cho nam",
-                                                    "slug": "khử-mùi-cho-nam",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.07.05_Kh%E1%BB%AD_m%C3%B9i_cho_nam_3.07.05-Mens_deodorant_Khu_mui_cho_nam.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDE5",
-                                    "name": "Chăm sóc răng miệng",
-                                    "slug": "chăm-sóc-răng-miệng",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.08_Ch%C4%83m_s%C3%B3c_r%C4%83ng_mi%E1%BB%87ng_3.08-Oral_Care_Cham_soc_rang_mieng.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDIw",
-                                                    "name": "Kem đánh răng",
-                                                    "slug": "kem-đánh-răng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.08.01_Kem_%C4%91%C3%A1nh_r%C4%83ng_3.08.01-Toothpaste_Kem_danh_rang.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDIx",
-                                                    "name": "Bàn chải đánh răng",
-                                                    "slug": "bàn-chải-đánh-răng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.08.02_B%C3%A0n_ch%E1%BA%A3i_%C4%91%C3%A1nh_r%C4%83ng_3.08.02-Toothbrush_Ban_chai_danh_rang.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDIy",
-                                                    "name": "Nước súc miệng",
-                                                    "slug": "nước-súc-miệng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.08.03_N%C6%B0%E1%BB%9Bc_s%C3%BAc_mi%E1%BB%87ng_3.08.03-Mouthwash_Nuoc_suc_mieng.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDIz",
-                                                    "name": "Sản phẩm làm trắng răng",
-                                                    "slug": "sản-phẩm-làm-trắng-răng",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDI0",
-                                                    "name": "Tăm chỉ nha khoa",
-                                                    "slug": "tăm-chỉ-nha-khoa",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.08.05_T%C4%83m_ch%E1%BB%89_nha_khoa_3.08.05-Dental_Floss__Pick_Tam_chi_nha_khoa.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDI1",
-                                                    "name": "Sản phẩm chăm sóc răng miệng khác",
-                                                    "slug": "sản-phẩm-chăm-sóc-răng-miệng-khác",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.08.06_S%E1%BA%A3n_ph%E1%BA%A9m_ch%C4%83m_s%C3%B3c_r%C4%83ng_mi%E1%BB%87ng_kh%C3%A1c_3.08.06-Denture_Care_San_pham_cham_soc_.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDI2",
-                                                    "name": "Bột đánh răng",
-                                                    "slug": "bột-đánh-răng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.08.07_B%E1%BB%99t_%C4%91%C3%A1nh_r%C4%83ng.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDI3",
-                                    "name": "Sản phẩm tẩy lông",
-                                    "slug": "sản-phẩm-tẩy-lông",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.09_S%E1%BA%A3n_ph%E1%BA%A9m_t%E1%BA%A9y_l%C3%B4ng_3.09-Deplilatories_San_pham_tay_long.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDI4",
-                                                    "name": "Kem tẩy lông",
-                                                    "slug": "kem-tẩy-lông",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.09.01_Kem_t%E1%BA%A9y_l%C3%B4ng_3.09.01-Hair_Removal_Cream_Kem_tay_long.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDI5",
-                                                    "name": "Sáp tẩy lông",
-                                                    "slug": "sáp-tẩy-lông",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDMw",
-                                                    "name": "Phụ kiện tẩy lông",
-                                                    "slug": "phụ-kiện-tẩy-lông",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.09.03_Ph%E1%BB%A5_ki%E1%BB%87n_t%E1%BA%A9y_l%C3%B4ng_3.09.03-Hair_Removal_Accessories_Phu_kien_tay_long.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            },
-            {
-                "node": {
-                    "id": "Q2F0ZWdvcnk6Mzk5",
-                    "name": "Sản phẩm tiện lợi",
-                    "slug": "sản-phẩm-tiện-lợi",
-                    "level": 0,
-                    "icon": {
-                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/05_S%E1%BA%A3n_ph%E1%BA%A9m_ti%E1%BB%87n_l%E1%BB%A3i_05-CVS_San_pham_tien_loi.png",
-                        "alt": ""
-                    },
-                    "backgroundImage": null,
-                    "translation": null,
-                    "metadata": [
-                        {
-                            "key": "color",
-                            "value": "#00B6F0"
-                        }
-                    ],
-                    "parent": null,
-                    "children": {
-                        "edges": [
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDU0",
-                                    "name": "Hàng tổng hợp",
-                                    "slug": "hàng-tổng-hợp",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.01_H%C3%A0ng_t%E1%BB%95ng_h%E1%BB%A3p_5.01-General_Merchandise_Hang_tong_hop.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDU1",
-                                                    "name": "Hàng dành cho du lịch",
-                                                    "slug": "hàng-dành-cho-du-lịch",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.01.05_H%C3%A0ng_d%C3%A0nh_cho_du_l%E1%BB%8Bch_5.01.05-Travel_Needs_Hang_danh_cho_du_lich.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDU2",
-                                                    "name": "Khăn giấy, khăn ướt",
-                                                    "slug": "khăn-giấy-khăn-ướt",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.01.06_Kh%C4%83n_gi%E1%BA%A5y_kh%C4%83n_%C6%B0%E1%BB%9Bt_5.01.06-Tissues_Khan_giay_khan_uot.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDU3",
-                                                    "name": "Hàng theo mùa",
-                                                    "slug": "hàng-theo-mùa",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.01.07_H%C3%A0ng_theo_m%C3%B9a_5.01.07-Seasonal_Items_Hang_theo_mua.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDU4",
-                                                    "name": "Thiết bị điện gia dụng",
-                                                    "slug": "home-appliance-thiet-bi-dien-gia-dung",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDU5",
-                                    "name": "Hàng bách hóa",
-                                    "slug": "hàng-bách-hóa",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/H%C3%A0ng_b%C3%A1ch_h%C3%B3a.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDYw",
-                                                    "name": "Thức uống",
-                                                    "slug": "thức-uống",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.02.01_Th%E1%BB%A9c_u%E1%BB%91ng_5.02.01-Beverage_Thuc_uong.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDYx",
-                                                    "name": "Bánh quy",
-                                                    "slug": "bánh-quy",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.02.02_B%C3%A1nh_quy_5.02.02-Biscuit_Cookies_Banh_quy.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDYy",
-                                                    "name": "Bánh snack",
-                                                    "slug": "bánh-snack",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.02.03_B%C3%A1nh_snack_5.02.03-Snacks_Banh_snack.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDYz",
-                                                    "name": "Kẹo, gum",
-                                                    "slug": "kẹo-gum",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.02.04_K%E1%BA%B9o_gum_5.02.04-Confectionery_Keo_gum.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDY0",
-                                                    "name": "Kem",
-                                                    "slug": "kem",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDY1",
-                                                    "name": "Sản phẩm sữa và từ sữa",
-                                                    "slug": "sản-phẩm-sữa-và-từ-sữa",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.02.06_S%E1%BA%A3n_ph%E1%BA%A9m_s%E1%BB%AFa_v%C3%A0_t%E1%BB%AB_s%E1%BB%AFa_5.02.06-Dairy_products_San_pham_sua_va_tu_sua.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDY2",
-                                                    "name": "Thực phẩm khác",
-                                                    "slug": "thực-phẩm-khác",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/5.02.07_Th%E1%BB%B1c_ph%E1%BA%A9m_kh%C3%A1c_5.02.07-Other_food_Thuc_pham_khac.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            },
-            {
-                "node": {
-                    "id": "Q2F0ZWdvcnk6NDAw",
-                    "name": "Thực phẩm chức năng",
-                    "slug": "thực-phẩm-chức-năng",
-                    "level": 0,
-                    "icon": {
-                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/06_Th%E1%BB%B1c_ph%E1%BA%A9m_ch%E1%BB%A9c_n%C4%83ng_06_06-Vitamin__Supplement_Thuc_pham_chuc_nang.png",
-                        "alt": ""
-                    },
-                    "backgroundImage": null,
-                    "translation": null,
-                    "metadata": [
-                        {
-                            "key": "backgroundColor",
-                            "value": "#05B085"
-                        }
-                    ],
-                    "parent": null,
-                    "children": {
-                        "edges": [
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDY3",
-                                    "name": "TPCN Nhóm tiêu hóa",
-                                    "slug": "tpcn-nhóm-tiêu-hóa",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.01_6.01-TPCN_Nh%C3%B3m_ti%C3%AAu_h%C3%B3a_6.01-VMS_For_Stomach_TPCN_Nhom_tieu_hoa.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDY4",
-                                                    "name": "TPCN cho Dạ Dày",
-                                                    "slug": "curcumin-tpcn-cho-da-day",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.01.01_6.01.01-Curcumin_TPCN_cho_Da_Day_6.01.01-Curcumin_TPCN_cho_Da_Day.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDY5",
-                                                    "name": "TPCN Nhóm tiêu hóa",
-                                                    "slug": "digestion-tpcn-nhom-tieu-hoa",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.01.02_6.01.02-Digestion_TPCN_Nhom_Tieu_Hoa_6.01.02-Digestion_TPCN_Nhom_Tieu_Hoa.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDcw",
-                                                    "name": "TPCN nhóm khác",
-                                                    "slug": "others-tpcn-khac",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDcx",
-                                    "name": "TPCN Nhóm tim mạch",
-                                    "slug": "tpcn-nhóm-tim-mạch",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.02_6.02-TPCN_Nh%C3%B3m_tim_m%E1%BA%A1ch_-_Huy%E1%BA%BFt_%C3%A1p_6.02-VMS_For_Heart_TPCN_Nhom_tim_mach_-_h.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDcy",
-                                                    "name": "TPCN Cho Huyết Áp",
-                                                    "slug": "blood-pressure-tpcn-cho-huyet-ap",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.02.01_6.02.01-Blood_Pressure_TPCN_Cho_Huyet_Ap_6.02.01-Blood_Pressure_TPCN_Cho_.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDcz",
-                                                    "name": "TPCN Tuần hoàn máu",
-                                                    "slug": "tpcn-tuần-hoàn-máu",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.02.02_6.02.02-TPCN_Tu%E1%BA%A7n_ho%C3%A0n_m%C3%A1u_6.02.02-Circulation_TPCN_Tuan_Hoan_Mau.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDc0",
-                                    "name": "TPCN Nhóm đường huyết",
-                                    "slug": "tpcn-nhóm-đường-huyết",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.02.03_6.02.03-TPCN_Nh%C3%B3m_%C4%91%C6%B0%E1%BB%9Dng_huy%E1%BA%BFt_6.02.03-Diabetic_TPCN_Nhom_Duong_Huyet.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": []
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDc1",
-                                    "name": "TPCN Nhóm hô hấp",
-                                    "slug": "tpcn-nhóm-hô-hấp",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.03_6.03-TPCN_Nh%C3%B3m_h%C3%B4_h%E1%BA%A5p_6.03-VMS_For_Lungs_TPCN_Nhom_ho_hap.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDc2",
-                                                    "name": "TPCN Nhóm Keo Ong",
-                                                    "slug": "propolis-tpcn-nhom-keo-ong",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.03.01_6.03.01-Propolis_TPCN_Nhom_Keo_Ong_6.03.01-Propolis_TPCN_Nhom_Keo_Ong.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDc3",
-                                                    "name": "Nhóm xịt họng thảo dược",
-                                                    "slug": "nhóm-xịt-họng-thảo-dược",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.03.02_6.03.02-Nh%C3%B3m_x%E1%BB%8Bt_h%E1%BB%8Dng_th%E1%BA%A3o_d%C6%B0%E1%BB%A3c_6.03.02-Herbal_Cough_TPCN_Nhom_Xit_Hong_T.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDc4",
-                                                    "name": "TPCN Cho Phổi Khác",
-                                                    "slug": "others-tpcn-cho-phoi-khac",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDc5",
-                                    "name": "TPCN Nhóm thần kinh",
-                                    "slug": "tpcn-nhóm-thần-kinh",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.04_6.04-TPCN_Nh%C3%B3m_th%E1%BA%A7n_kinh_6.04-VMS_For_Brain_TPCN_Nhom_than_kinh.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDgw",
-                                                    "name": "TPCN giảm căng thẳng",
-                                                    "slug": "anti-stress-tpcn-giam-cang-thang",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.04.01_6.04.01-Anti_Stress_TPCN_Giam_Cang_Thang_6.04.01-Anti_Stress_TPCN_Giam_Ca.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDgx",
-                                                    "name": "TPVN Giúp trí nhớ",
-                                                    "slug": "giúp-trí-nhớ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.04.02_6.04.02-Gi%C3%BAp_tr%C3%AD_nh%E1%BB%9B_6.04.02-Memory_TPCN_Giup_Tri_Nho.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDgy",
-                                                    "name": "TPCN tuần hoàn máu",
-                                                    "slug": "circulation-tpcn-tuan-hoan-nao",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.04.03_6.04.03-Circulation_TPCN_Tuan_Hoan_Nao_6.04.03-Circulation_TPCN_Tuan_Hoan.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDgz",
-                                                    "name": "TPCN hỗ trợ giấc ngủ",
-                                                    "slug": "sleep-tpcn-htro-gngu",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.04.04_6.04.04-Sleep_TPCN_H.Tro_G.Ngu_6.04.04-Sleep_TPCN_H.Tro_G.Ngu.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDg0",
-                                    "name": "TPCN Nhóm cơ xương khớp",
-                                    "slug": "tpcn-nhóm-cơ-xương-khớp",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.05_6.05-TPCN_Nh%C3%B3m_c%C6%A1_x%C6%B0%C6%A1ng_kh%E1%BB%9Bp_6.05-VMS_For_Bone__Joint_TPCN_Nhom_co_xuong_kho.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDg1",
-                                                    "name": "TPCN chống loãng xương",
-                                                    "slug": "tpcn-chống-loãng-xương",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.05.01_6.05.01-TPCN_ch%E1%BB%91ng_lo%C3%A3ng_x%C6%B0%C6%A1ng_6.05.01-Bone_Health_TPCN_Chong_Loang_Xuong.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDg2",
-                                                    "name": "TPCN hỗ trợ khớp",
-                                                    "slug": "tpcn-hỗ-trợ-khớp",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.05.02_6.05.02-TPCN_h%E1%BB%97_tr%E1%BB%A3_kh%E1%BB%9Bp_6.05.02-Joint_Health_TPCN_H.Tro_Khop.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDg3",
-                                                    "name": "TPCN hỗ trợ gút",
-                                                    "slug": "tpcn-hỗ-trợ-gút",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.05.03_6.05.03-TPCN_h%E1%BB%97_tr%E1%BB%A3_g%C3%BAt_6.05.03-Gout_Health_TPCN_H.Tro_Gut.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDg4",
-                                                    "name": "TPCN hỗ trợ xương khớp khác",
-                                                    "slug": "tpcn-hỗ-trợ-xương-khớp-khác",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDg5",
-                                    "name": "TPCN Giảm cân",
-                                    "slug": "tpcn-giảm-cân",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.06_6.06-TPCN_Gi%E1%BA%A3m_c%C3%A2n_6.06-VMS_Dietary_TPCN_Giam_Can.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDkw",
-                                                    "name": "TPCN đốt mỡ",
-                                                    "slug": "tpcn-đốt-mỡ",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDkx",
-                                                    "name": "TPCN hỗ trợ chuyển hóa mỡ",
-                                                    "slug": "tpcn-hỗ-trợ-chuyển-hóa-mỡ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.06.02_6.06.02-TPCN_h%E1%BB%97_tr%E1%BB%A3_chuy%E1%BB%83n_h%C3%B3a_m%E1%BB%A1_6.06.02-Dietary_Metabolism_TPCN_H.Tro_C.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDky",
-                                                    "name": "TPCN tăng cân",
-                                                    "slug": "tpcn-tăng-cân",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.06.03_6.06.03-TPCN_t%C4%83ng_c%C3%A2n_6.06.03-Weight_Gain_TPCN_Tang_Can.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDkz",
-                                    "name": "TPCN Chăm sóc sắc đẹp",
-                                    "slug": "chăm-sóc-sắc-đẹp-2",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.17.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDk0",
-                                                    "name": "TPCN đẹp da",
-                                                    "slug": "tpcn-đẹp-da",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDk1",
-                                                    "name": "TPCN cho tóc",
-                                                    "slug": "tpcn-cho-tóc",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDk2",
-                                                    "name": "TPCN hỗ trợ da móng tóc",
-                                                    "slug": "tpcn-hỗ-trợ-da-móng-tóc",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NDk3",
-                                    "name": "TPCN Chăm sóc sức khỏe nam và nữ",
-                                    "slug": "tpcn-chăm-sóc-sức-khỏe-nam-và-nữ",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.08_6.08-TPCN_Ch%C4%83m_s%C3%B3c_s%E1%BB%A9c_kh%E1%BB%8Fe_nam_v%C3%A0_n%E1%BB%AF_6.08-VMS_Mens__Womens_Health_TPCN_Cham.jpg",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDk4",
-                                                    "name": "TPCN hỗ trợ sinh lý nam",
-                                                    "slug": "tpcn-hỗ-trợ-sinh-lý-nam",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.08.01_6.08.01-_TPCN_h%E1%BB%97_tr%E1%BB%A3_sinh_l%C3%BD_nam_6.08.01-_Men_Health_TPCN_H.Tro_S.ly_Nam.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NDk5",
-                                                    "name": "TPCN hỗ trợ sinh lý nữ",
-                                                    "slug": "tpcn-hỗ-trợ-sinh-lý-nữ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.08.02_6.08.02-TPCN_h%E1%BB%97_tr%E1%BB%A3_sinh_l%C3%BD_n%E1%BB%AF_6.08.02-Women_Health_TPCN_H.Tro_S.Ly_Nu.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTAw",
-                                    "name": "TPCN Nhóm Mắt/Tai/Mũi",
-                                    "slug": "tpcn-nhóm-mắttaimũi",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.09_6.09-TPCN_Nh%C3%B3m_M%E1%BA%AFtTaiM%C5%A9i_6.09-VMS_For_EyeEarNose_TPCN_Nhom_MatTaiMui_-_Copy.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTAx",
-                                                    "name": "TPCN cho mắt",
-                                                    "slug": "tpcn-cho-mắt",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.09.01_6.09.01-TPCN_cho_m%E1%BA%AFt_6.09.01-Eye_Health_TPCN_Cho_Mat.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTAy",
-                                                    "name": "TPCN cho tai",
-                                                    "slug": "tpcn-cho-tai",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTAz",
-                                                    "name": "TPCN cho mũi",
-                                                    "slug": "tpcn-cho-mũi",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTA0",
-                                    "name": "TPCN Vitamin tổng hợp và khoáng chất",
-                                    "slug": "tpcn-vitamin-tổng-hợp-và-khoáng-chất",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.10_6.10-TPCN_Vitamin_t%E1%BB%95ng_h%E1%BB%A3p_v%C3%A0_kho%C3%A1ng_ch%E1%BA%A5t_6.10-VMS_General_Vitamins__Mineral.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTA1",
-                                                    "name": "Vitamin tổng hợp",
-                                                    "slug": "vitamin-tổng-hợp",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.10.01_6.10.01-Vitamin_t%E1%BB%95ng_h%E1%BB%A3p_6.10.01-Multivitamin_TPCN_Nhom_Vitamin_Tong_Hop.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTA3",
-                                                    "name": "Vitamin C",
-                                                    "slug": "vitamin-c",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.10.02_6.10.02-Vitamin_C_6.10.02-Vitamin_C_TPCN_VitaminC.jpg",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTA4",
-                                                    "name": "Vitamin A/B/D",
-                                                    "slug": "vitamin-abd",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.10.03_6.10.03-Vitamin_ABD_6.10.03-Herbal_TPCN_Thao_Duoc.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTA5",
-                                                    "name": "TPCN khác",
-                                                    "slug": "tpcn-khác",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTEw",
-                                                    "name": "TPCN cung cấp sắt",
-                                                    "slug": "iron-tpcn-cung-cap-sat",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTEx",
-                                                    "name": "TPCN nhóm Omega",
-                                                    "slug": "omegas-tpcn-nhom-omega",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.02.04_6.02.04-Omegas_TPCN_Nhom_Omega_6.02.04-Omegas_TPCN_Nhom_Omega.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTEy",
-                                    "name": "TPCN Chăm sóc tóc",
-                                    "slug": "tpcn-chăm-sóc-tóc",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.11.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": []
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTEz",
-                                    "name": "TPCN Nhóm khác",
-                                    "slug": "tpcn-nhóm-khác",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.14_6.14-TPCN_Nh%C3%B3m_kh%C3%A1c_6.14-VMS_Others_TPCN_Nhom_khac.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": []
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTE0",
-                                    "name": "TPCN cho gan",
-                                    "slug": "tpcn-cho-gan",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.15_6.15-TPCN_cho_gan_6.15-VMS_For_Liver_TPCN_Cho_Gan.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTE1",
-                                                    "name": "TPCN giải độc gan",
-                                                    "slug": "tpcn-giải-độc-gan",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTE2",
-                                                    "name": "TPCN dành cho gan khác",
-                                                    "slug": "tpcn-dành-cho-gan-khác",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            },
-            {
-                "node": {
-                    "id": "Q2F0ZWdvcnk6NDAx",
-                    "name": "Mẹ và Bé",
-                    "slug": "mẹ-và-bé",
-                    "level": 0,
-                    "icon": {
-                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.12.png",
-                        "alt": ""
-                    },
-                    "backgroundImage": null,
-                    "translation": null,
-                    "metadata": [
-                        {
-                            "key": "color",
-                            "value": "#0F62F9"
-                        }
-                    ],
-                    "parent": null,
-                    "children": {
-                        "edges": [
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTIw",
-                                    "name": "Chăm sóc em bé",
-                                    "slug": "chăm-sóc-em-bé",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/photo_2021-08-23_21-10-35_8080.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTIx",
-                                                    "name": "Phụ kiện cho bé",
-                                                    "slug": "phụ-kiện-cho-bé",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.01_Ph%E1%BB%A5_ki%E1%BB%87n_cho_b%C3%A9_3.01.01-Baby_Accessories_Phu_kien_cho_be.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTIy",
-                                                    "name": "Kem & dầu dưỡng cho bé",
-                                                    "slug": "kem-dầu-dưỡng-cho-bé",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.02_Kem__d%E1%BA%A7u_d%C6%B0%E1%BB%A1ng_cho_b%C3%A9_3.01.02-Baby_Cream__Oil_Kem__dau_duong_cho_be.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTIz",
-                                                    "name": "Phấn rôm",
-                                                    "slug": "phấn-rôm",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.03_Ph%E1%BA%A5n_r%C3%B4m_3.01.03-Talcum_Powder_Phan_rom.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTI0",
-                                                    "name": "Đồ dùng phòng tắm cho bé",
-                                                    "slug": "đồ-dùng-phòng-tắm-cho-bé",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTI1",
-                                                    "name": "Dầu gội cho bé",
-                                                    "slug": "dầu-gội-cho-bé",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.06_D%E1%BA%A7u_g%E1%BB%99i_cho_b%C3%A9_3.01.06-Baby_Shampoo_Dau_goi_cho_be.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTI2",
-                                                    "name": "Tã cho bé",
-                                                    "slug": "tã-cho-bé",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.07_T%C3%A3_cho_b%C3%A9_3.01.07-Baby_Diaper_Ta_cho_be.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTI3",
-                                                    "name": "Sản phẩm chống muỗi cho bé",
-                                                    "slug": "sản-phẩm-chống-muỗi-cho-bé",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.08_S%E1%BA%A3n_ph%E1%BA%A9m_ch%E1%BB%91ng_mu%E1%BB%97i_cho_b%C3%A9_3.01.08-Insect_Repellent_San_pham_chong_muoi_c.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTI4",
-                                                    "name": "Khăn ướt cho bé",
-                                                    "slug": "khăn-ướt-cho-bé",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.09_Kh%C4%83n_%C6%B0%E1%BB%9Bt_cho_b%C3%A9_3.01.09-Baby_Wipes_Khan_uot_cho_be.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTI5",
-                                                    "name": "Sản phẩm chăm sóc sức khoẻ cho bé",
-                                                    "slug": "sản-phẩm-chăm-sóc-sức-khoẻ-cho-bé",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.10_S%E1%BA%A3n_ph%E1%BA%A9m_ch%C4%83m_s%C3%B3c_s%E1%BB%A9c_kho%E1%BA%BB_cho_b%C3%A9_3.01.10-Baby_Healthcare_Products_San_ph.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTMw",
-                                                    "name": "Thực phẩm dành cho bé",
-                                                    "slug": "thực-phẩm-dành-cho-bé",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.11_Th%E1%BB%B1c_ph%E1%BA%A9m_d%C3%A0nh_cho_b%C3%A9_3.01.11-Baby_Food_Thuc_pham_danh_cho_be.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTMx",
-                                                    "name": " Sữa dành cho bé",
-                                                    "slug": "sữa-dành-cho-bé",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.12_S%E1%BB%AFa_d%C3%A0nh_cho_b%C3%A9_3.01.12-Baby_Milk_Sua_danh_cho_be.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTMy",
-                                                    "name": "Chăm sóc cá nhân cho bé",
-                                                    "slug": "baby-toiletries-cham-soc-ca-nhan-cho-be",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.13_3.01.13-Baby_Toiletries_Cham_soc_ca_nhan_cho_be_3.01.13-Baby_Toiletries_C.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTMz",
-                                                    "name": "Tăm bông cho bé",
-                                                    "slug": "baby-cotton-tam-bong-cho-be",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.01.14_3.01.14-Baby_Cotton_Tam_bong_cho_be_3.01.14-Baby_Cotton_Tam_bong_cho_be.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTM0",
-                                    "name": "TPCN dành cho trẻ em",
-                                    "slug": "tpcn-dành-cho-trẻ-em",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.12_6.12-TPCN_D%C3%A0nh_cho_tr%E1%BA%BB_em_6.12-VMS_For_Kids_TPCN_Danh_cho_tre_em.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTM1",
-                                                    "name": "TPCN cho não của trẻ",
-                                                    "slug": "tpcn-cho-não-của-trẻ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.12.01_6.12.01-TPCN_cho_n%C3%A3o_c%E1%BB%A7a_tr%E1%BA%BB_6.12.01-Brain_Health_TPCN_Cho_Nao_Tre_Em.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTM2",
-                                                    "name": "TPCN tiêu hóa cho trẻ",
-                                                    "slug": "tpcn-tiêu-hóa-cho-trẻ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.12.02_6.12.02-TPCN_ti%C3%AAu_h%C3%B3a_cho_tr%E1%BA%BB_6.12.02-Digestion_TPCN_Tieu_Hoa_Cho_Tre_Em.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTM3",
-                                                    "name": "TPCN dành cho mắt của trẻ",
-                                                    "slug": "tpcn-dành-cho-mắt-của-trẻ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.12.03_6.12.03-TPCN_d%C3%A0nh_cho_m%E1%BA%AFt_c%E1%BB%A7a_tr%E1%BA%BB_6.12.03-Kid_Herbal_Cough_TPCN_Thao_Duoc.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTM4",
-                                                    "name": "TPCN tăng chiều cao cho trẻ",
-                                                    "slug": "tpcn-tăng-chiều-cao-cho-trẻ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.12.04_6.12.04-TPCN_t%C4%83ng_chi%E1%BB%81u_cao_cho_tr%E1%BA%BB_6.12.04-Kid_Growth_TPCN_Phat_Trien_Ch.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTM5",
-                                                    "name": "TPCN tổng hợp cho trẻ",
-                                                    "slug": "tpcn-tổng-hợp-cho-trẻ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.12.05_6.12.05-TPCN_t%E1%BB%95ng_h%E1%BB%A3p_cho_tr%E1%BA%BB_6.12.05-Multivitamin_TPCN_T.Hop_Cho_Tre.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTQw",
-                                                    "name": "TPCN tăng cân cho trẻ",
-                                                    "slug": "tpcn-tăng-cân-cho-trẻ",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.12.06_6.12.06-TPCN_t%C4%83ng_c%C3%A2n_cho_tr%E1%BA%BB_6.12.06-Weight_Gain_TPCN_Tang_Can_Cho_Tre.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTQx",
-                                                    "name": "TPCN khác cho trẻ",
-                                                    "slug": "tpcn-khác-cho-trẻ",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTQy",
-                                                    "name": "TPCN Cung cấp vit D cho bé",
-                                                    "slug": "vitamin-d-tpcn-cung-cap-vit-d-cho-be",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.12.08_6.12.08-_Vitamin_D_TPCN_Cung_Cap_Vit_D_Cho_Be_6.12.08-_Vitamin_D_TPCN_Cun.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTQz",
-                                    "name": "Sản phẩm dành cho mẹ",
-                                    "slug": "mom-care-san-pham-danh-cho-me",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/photo_2021-08-23_21-08-20.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTQ0",
-                                                    "name": "Dầu dưỡng thể cho mẹ",
-                                                    "slug": "body-massage-oil-for-mom-dau-duong-the-cho-me",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.10.01_3.10.01-Body__Massage_Oil_For_Mom_Dau_duong_the_cho_Me_3.10.01-Body__Mass.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTQ1",
-                                                    "name": "Sữa cho mẹ",
-                                                    "slug": "milk-for-mom-sua-cho-me",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.10.02_3.10.02-Milk_For_Mom_Sua_cho_Me_3.10.02-Milk_For_Mom_Sua_cho_Me.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTQ2",
-                                                    "name": "Phụ kiện cho mẹ",
-                                                    "slug": "accessories-for-mom-phu-kien-cho-me",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.10.03_3.10.03-Accessories_For_Mom_3.10.03-Accessories_For_Mom.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTQ3",
-                                                    "name": "Băng lót cho mẹ",
-                                                    "slug": "sanitary-napkins-for-mom-bang-lot-cho-me",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.10.04_3.10.04-Sanitary_Napkins_For_Mom_Sanitary_Napkins_For_Mom_3.10.04-Sanitar.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTQ4",
-                                                    "name": "Dụng cụ chăm sóc cho mẹ",
-                                                    "slug": "healthcare-equipment-for-mom-dung-cu-cham-soc-cho-me",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.10.05_3.10.05-Healthcare_Equipment_For_Mom_Measurement_of_Care_for_Me_3.10.05-H.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTQ5",
-                                                    "name": "Dung dịch cho mẹ",
-                                                    "slug": "feminine-wash-for-mom-dung-dich-cho-me",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.10.06_3.10.06-Feminine_Wash_For_Mom_Solution_for_Me_3.10.06-Feminine_Wash_For_M.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTUw",
-                                                    "name": "Thực phẩm cho mẹ",
-                                                    "slug": "food-for-mom-thuc-pham-cho-me",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/3.10.07_3.10.07-Food_For_Mom_Food_For_Mom_3.10.07-Food_For_Mom_Food_For_Mom.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTUx",
-                                                    "name": "Chăm sóc da mặt cho mẹ",
-                                                    "slug": "face-care-for-mom-cham-soc-da-mat-cho-me",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTUy",
-                                    "name": "TPCN dành cho phụ nữ mang thai",
-                                    "slug": "tpcn-dành-cho-phụ-nữ-mang-thai",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/6.16.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": []
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            },
-            {
-                "node": {
-                    "id": "Q2F0ZWdvcnk6NTU1",
-                    "name": "Chăm sóc sắc đẹp",
-                    "slug": "chăm-sóc-sắc-đẹp",
-                    "level": 0,
-                    "icon": {
-                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/04_Ch%C4%83m_s%C3%B3c_s%E1%BA%AFc_%C4%91%E1%BA%B9p_04-Beauty_Care_Cham_soc_sac_dep.png",
-                        "alt": ""
-                    },
-                    "backgroundImage": null,
-                    "translation": null,
-                    "metadata": [
-                        {
-                            "key": "color",
-                            "value": "#5DAC45"
-                        }
-                    ],
-                    "parent": null,
-                    "children": {
-                        "edges": [
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTU2",
-                                    "name": "Chăm sóc mặt",
-                                    "slug": "chăm-sóc-mặt",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.01_Ch%C4%83m_s%C3%B3c_m%E1%BA%B7t_4.01-Face_Care_Cham_soc_mat.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTU3",
-                                                    "name": "Nước tẩy trang",
-                                                    "slug": "nước-tẩy-trang",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.01.01_N%C6%B0%E1%BB%9Bc_t%E1%BA%A9y_trang_4.01.01-Makeup_Remover_Nuoc_tay_trang.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTU4",
-                                                    "name": "Mặt nạ dưỡng da",
-                                                    "slug": "mặt-nạ-dưỡng-da",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.01.02_M%E1%BA%B7t_n%E1%BA%A1_d%C6%B0%E1%BB%A1ng_da_4.01.02-Beauty_Facemask_Mat_na_duong_da.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTU5",
-                                                    "name": "Sữa rửa mặt",
-                                                    "slug": "sữa-rửa-mặt",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.01.03_S%E1%BB%AFa_r%E1%BB%ADa_m%E1%BA%B7t_4.01.03-Facial_Cleanser_Sua_rua_mat.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTYw",
-                                                    "name": "Tẩy tế bào chết cho mặt",
-                                                    "slug": "tẩy-tế-bào-chết-cho-mặt",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.01.04_T%E1%BA%A9y_t%E1%BA%BF_b%C3%A0o_ch%E1%BA%BFt_cho_m%E1%BA%B7t_4.01.04-Face_Scrubs__Exfoliators_Tay_te_bao_chet_.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTYx",
-                                                    "name": "Dưỡng môi",
-                                                    "slug": "dưỡng-môi",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.01.05_D%C6%B0%E1%BB%A1ng_m%C3%B4i_4.01.05-Lip_Balm__Treatment_Duong_moi.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTYy",
-                                                    "name": "Kem dưỡng ẩm và dưỡng da",
-                                                    "slug": "kem-dưỡng-ẩm-và-dưỡng-da",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.01.06_Kem_d%C6%B0%E1%BB%A1ng_%E1%BA%A9m_v%C3%A0_d%C6%B0%E1%BB%A1ng_da_4.01.06-Moisturizers__Creams_Kem_duong_am__duong.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTYz",
-                                                    "name": "Nước hoa hồng & Xịt khoáng",
-                                                    "slug": "nước-hoa-hồng-xịt-khoáng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.01.07_N%C6%B0%E1%BB%9Bc_hoa_h%E1%BB%93ng__X%E1%BB%8Bt_kho%C3%A1ng_4.01.07-Toner__Mist_Nuoc_hoa_hong__xit_khoang.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTY0",
-                                                    "name": "Kem trị mụn",
-                                                    "slug": "kem-trị-mụn",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.01.08_Kem_tr%E1%BB%8B_m%E1%BB%A5n_4.01.08-Acne_Treatment_Kem_tri_mun.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTY1",
-                                    "name": "Sản phẩm chống nắng",
-                                    "slug": "sản-phẩm-chống-nắng",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.02_S%E1%BA%A3n_ph%E1%BA%A9m_ch%E1%BB%91ng_n%E1%BA%AFng_4.02-Sun_Care_San_pham_chong_nang.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTY2",
-                                                    "name": "Kem chống nắng dành cho mặt",
-                                                    "slug": "kem-chống-nắng-dành-cho-mặt",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.02.01_Kem_ch%E1%BB%91ng_n%E1%BA%AFng_d%C3%A0nh_cho_m%E1%BA%B7t_4.02.01-For_Face_Kem_chong_nang_danh_cho_mat.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTY3",
-                                                    "name": "Kem chống nắng toàn thân",
-                                                    "slug": "kem-chống-nắng-toàn-thân-2",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.02.02_Kem_ch%E1%BB%91ng_n%E1%BA%AFng_to%C3%A0n_th%C3%A2n_4.02.02-For_Body_Kem_chong_nang_toan_than.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTY4",
-                                                    "name": "Kem chống nắng cho mặt và cơ thể",
-                                                    "slug": "kem-chống-nắng-cho-mặt-và-cơ-thể",
-                                                    "translation": null,
-                                                    "icon": null,
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTY5",
-                                                    "name": "Dưỡng da sau khi đi nắng",
-                                                    "slug": "dưỡng-da-sau-khi-đi-nắng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.02.04_D%C6%B0%E1%BB%A1ng_da_sau_khi_%C4%91i_n%E1%BA%AFng_4.02.04-Aftersun_Gel_Duong_da_sau_khi_di_nang.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            },
-                            {
-                                "node": {
-                                    "id": "Q2F0ZWdvcnk6NTcw",
-                                    "name": "Dụng cụ làm đẹp",
-                                    "slug": "dụng-cụ-làm-đẹp",
-                                    "translation": null,
-                                    "icon": {
-                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.03_D%E1%BB%A5ng_c%E1%BB%A5_l%C3%A0m_%C4%91%E1%BA%B9p_4.03-Beauty_Accessories_Dung_cu_lam_dep.png",
-                                        "alt": ""
-                                    },
-                                    "backgroundImage": null,
-                                    "metadata": [],
-                                    "children": {
-                                        "edges": [
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTcx",
-                                                    "name": "Cotton",
-                                                    "slug": "cotton",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.03.01_Cotton_4.03.01-Cotton_Cotton.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTcy",
-                                                    "name": "Phụ kiện trang điểm",
-                                                    "slug": "phụ-kiện-trang-điểm",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.03.02_Ph%E1%BB%A5_ki%E1%BB%87n_trang_%C4%91i%E1%BB%83m_4.03.02-Makeup_Accessories_Phu_kien_trang_diem.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTcz",
-                                                    "name": "Phụ kiện tóc",
-                                                    "slug": "phụ-kiện-tóc",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.03.03_Ph%E1%BB%A5_ki%E1%BB%87n_t%C3%B3c_4.03.03-Hair_Accessories_Phu_kien_toc.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            },
-                                            {
-                                                "node": {
-                                                    "id": "Q2F0ZWdvcnk6NTc0",
-                                                    "name": "Phụ kiện làm móng",
-                                                    "slug": "phụ-kiện-làm-móng",
-                                                    "translation": null,
-                                                    "icon": {
-                                                        "url": "https://pharmacity-ecm-core-dev.s3.amazonaws.com/category-icons/4.03.04_Ph%E1%BB%A5_ki%E1%BB%87n_l%C3%A0m_m%C3%B3ng_4.03.04-Nail_Accessories_Phu_kien_lam_mong.png",
-                                                        "alt": ""
-                                                    },
-                                                    "backgroundImage": null,
-                                                    "metadata": [],
-                                                    "children": {
-                                                        "edges": []
-                                                    }
-                                                }
-                                            }
-                                        ]
-                                    }
-                                }
-                            }
-                        ]
-                    }
-                }
-            }
-        ],
-        "pageInfo": {
-            "endCursor": "WyI1NTUiXQ==",
-            "hasNextPage": false,
-            "hasPreviousPage": false,
-            "startCursor": "WyIyODYiXQ=="
-        }
-    }
-}
