--- components/Layout/Header/data/index.js
@@ -1,1 +1,0 @@
-export * from './dataDropdown'
