+++ components/Layout/Header/scss/Admin.module.scss
@@ -1,7 +1,5 @@
-.ant-layout {
-    margin-top: 80px;
+.ant-back {
     box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
     padding: 10px 0;
-    background: white;
 }
