--- components/Layout/Header/scss/Dropdown.module.scss
@@ -1,78 +1,0 @@
-.dropdown-box-center {
-    width: 50%;
-    display: flex;
-    flex-wrap: wrap;
-    padding: 5px 5px;
-
-    .item-box {
-        cursor: pointer;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        text-align: center;
-        flex-direction: column;
-        width: 20%;
-        height: 100px;
-        margin: 0 0 5px;
-        padding: 0 10px;
-
-        .item-name {
-            height: 15px;
-            font-size: 14px;
-        }
-
-        img {
-            height: 60px;
-            width: 60px;
-        }
-    }
-}
-
-.dropdown-box {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    width: 23%;
-}
-
-.dropdown-box:first-child::before {
-    position: absolute;
-    content: "";
-    top: -20px;
-    right: 0;
-    height: 2.5rem;
-    background-color: transparent;
-    border-left: 15px solid transparent;
-    border-right: 15px solid transparent;
-    border-bottom: 25px solid white;
-
-    transform: translateY(-100%);
-}
-
-.dropdown-item {
-    height: 50px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    cursor: pointer;
-    color: black;
-    margin: 0 0 5px;
-    padding-left: 10px;
-
-    svg {
-        path {
-            stroke: black;
-        }
-    }
-
-    &.active {
-        background-color: #0f62f9;
-        color: white;
-        border-radius: 12px;
-        svg {
-            path {
-                stroke: white;
-            }
-        }
-    }
-}
