+++ components/Layout/index.js
@@ -1,12 +1,11 @@
 import React from 'react';
-import Header from "./Header";
+import Header from "./Header"
 import Footer from "./Footer";
 
 const Layout = ({ children }) => {
     return (
         <>
-            {/* <Header /> */}
-            {children}
+            <Header children={children} />
             {/* <Footer/> */}
         </>
     )
